Change license to LGPL-2.0-or-later and GPL-2.0-or-later
Currently libimaevm provided by this project is used by the tool evmctl,
which is also provided by this project.

An issue was reported about using libimaevm with other software. Its
GPL2-only license makes it incompatible to use with other licenses, in
particular GPL3-only.

To address this issue, change the project license to GPL-2.0-or-later
and libimaevm to LGPL 2.0 or later.

Signed-off-by: Dmitry Kasatkin <dmitry.kasatkin@gmail.com>